fix: drop shipped JSX global augmentation; tighten lint relaxations
Remove the `declare global { namespace JSX }` block from src/types.ts: it
shipped as `type Element = Node` (an alias, not a mergeable interface) and
would collide with a consumer's own JSX.Element. The JSX factory binding
belongs in the actual JSX consumer (markup). Tests still use JSX, so the
binding now lives test-only in __tests__/jsx.d.ts (outside the src build).

Tighten the lint relaxations accordingly: drop no-namespace (only the JSX
block needed it); replace the blanket no-explicit-any off with site-local
disables on the two documented compose/sequence signatures so new
undocumented any is caught; scope naming-convention to allow leading
underscores rather than disabling it wholesale.
